feat(controllers/Categoryr): add endpoint POST /categories
